fix(trusty-mpm): correct banner two_panel right-column no-inner-border rendering
The right_col_no_inner_border test panicked because
render_two_panel_banner() loaded banner art from ~/.trusty-mpm/banner.txt
unconditionally, coupling every test that exercised it to whatever the
operator's own machine happened to have on disk. On a machine with a
customised banner containing legitimate rounded box-drawing glyphs
(robot-face art using ╭ ╮ ╰ ╯), the test's assertion that content rows
never contain inner-border characters failed — not because rendering was
broken, but because the test had no control over its own input.

Split render_two_panel_banner into a thin disk-backed wrapper plus a new
pure render_two_panel_banner_with_image(data, term_width, reconnecting,
image_lines, image_cols) that accepts the art explicitly. Production
callers are unaffected (render_two_panel_banner keeps the exact same
signature and disk-loading behavior); every two_panel test now calls the
image-parameterized function with a fixed, deterministic art source
(shade_image(DEFAULT_BANNER_ART, false)), making the suite hermetic and
independent of $HOME state.

Also splits two_panel.rs into a two_panel/ module (mod.rs + tests.rs) to
keep production code under the workspace's 500-SLOC cap after the fix,
moving the test suite into its own file under the 1500-SLOC test-file cap.

Closes #2224
